Archivos modificados, version final subida

+++ Funciones.h
@@ -1,57 +1,52 @@
 
-/** \brief Pide al usuario que ingrese un numero por teclado, tanto valor entero como valores con coma
+/** \brief Pide un valor al usuario
  *
- * \return double Nos devuelve el valor del numero ingresado por el usuario
+ * \return double devuelve un valor de tipo double
  *
  */
 double getNumber(); // Prototipo de la funcion que solicita los valores
 
-
 /** \brief Toma dos valores y los resta
  *
- * \param number1 double Primer valor a ser restado
- * \param number2 double Segundo valor
+ * \param number1 double Primer operador
+ * \param number2 double Segundo operador
  * \return double Devuelve el resultado de la resta
  *
  */
 double subtraction(double number1, double number2); // Prototipo de la funcion resta
 
-
 /** \brief Toma dos valores y los suma
  *
- * \param number1 double Primer valor a sumar
- * \param number2 double Segundo valor a sumar
+ * \param number1 double Primer operador
+ * \param number2 double Segundo operador
  * \return double Devuelve el resultado de la suma
  *
  */
 double sum(double number1, double number2); // Prototipo de la funcion suma
 
-
 /** \brief Toma dos valores y los multiplica
  *
- * \param number1 double Primer valor para la operacion
- * \param number2 double Segundo valor para la operacion
+ * \param number1 double Primer operador
+ * \param number2 double Segundo operador
  * \return double Devuelve el resultado de la multiplicacion
  *
  */
 double multiplication(double number1, double number2); // Prototipo de la funcion multiplicacion
 
-
 /** \brief Toma dos valores y los divide
  *
- * \param number1 double Primer valor, dividendo
- * \param number2 double Segundo valor, divisor
- * \return double Devuelve el resultado de la division
+ * \param number1 double Primer operador
+ * \param number2 double Segundo operador
+ * \return double Devuelve el valor de la division
  *
  */
 double division(double number1, double number2); // Prototipo de la funcion division
 
-
-/** \brief Toma dos valores y averigua el factorial de cada uno
+/** \brief Toma dos valores y averigua sus factoriales
  *
  * \param number1 double Primer valor
  * \param number2 double Segundo valor
- * \return double Devuelve el valor del factorial de ambos numeros
+ * \return long Devuelve el factorial
  *
  */
-double factorial(double number1, double number2); // Prototipo de la funcion de factorizacion
+long factorial(double number1, double number2); // Prototipo de la funcion de factorizacion
